refactor: timeTable 관련 쿼리 구조 변경
