+++ apps/client/src/pages/time-table/hooks/use-patch-time-table-mutation.ts
@@ -1,6 +1,6 @@
 import { useMutation, useQueryClient } from '@tanstack/react-query';
 
-import { patchFestivalTimetable } from '@shared/apis/time-table/festival-timetable';
+import { patchFestivalTimetable } from '@shared/apis/time-table/festival-time-table-queries';
 import { FESTIVAL_TIMETABLE_QUERY_KEY } from '@shared/constants/query-key';
 import { UserTimetable } from '@shared/types/timetable-response';
