+++ apps/client/src/shared/apis/time-table/festival-time-table-queries.ts
@@ -1,4 +1,8 @@
+import { queryOptions } from '@tanstack/react-query';
+
+import { del, get, patch } from '@shared/apis/config/instance';
 import { END_POINT } from '@shared/constants/api';
+import { FESTIVAL_TIMETABLE_QUERY_KEY } from '@shared/constants/query-key';
 import { BaseResponse } from '@shared/types/api';
 import {
   FestivalTimetableResponse,
@@ -7,13 +11,36 @@ import {
 } from '@shared/types/festival-timetable-response';
 import { UserTimetableResponse } from '@shared/types/timetable-response';
 
-import { del, get, patch } from '../config/instance';
+export const FESTIVAL_TIMETABLE_QUERY_OPTIONS = {
+  ALL: () =>
+    queryOptions({
+      queryKey: FESTIVAL_TIMETABLE_QUERY_KEY.ALL,
+    }),
+  ONBOARDING: () =>
+    queryOptions({
+      queryKey: FESTIVAL_TIMETABLE_QUERY_KEY.ONBOARDING(),
+      queryFn: getTimeTableCreationHistory,
+    }),
+  FESTIVAL_TIMETABLE: (festivalId: number) =>
+    queryOptions({
+      queryKey:
+        FESTIVAL_TIMETABLE_QUERY_KEY.DELETE_TIME_TABLE_FESTIVAL(festivalId),
+      queryFn: () => getFestivalTimetable(festivalId),
+    }),
+};
 
 export const getFestivalTimetable = async (
   festivalDateId: number,
 ): Promise<FestivalTimetableResponseExtended> => {
   const response = await get<BaseResponse<FestivalTimetableResponseExtended>>(
     END_POINT.GET_FESTIVAL_TIMETABLE(festivalDateId),
+  );
+  return response.data;
+};
+
+export const getTimeTableCreationHistory = async () => {
+  const response = await get<BaseResponse<TimeTableCreationHistory>>(
+    END_POINT.FETCH_TIMETABLE_CREATION_HISTORY,
   );
   return response.data;
 };
@@ -33,11 +60,4 @@ export const patchFestivalTimetable = async (
     END_POINT.GET_FESTIVAL_BUTTON,
     requestData,
   );
-};
-
-export const getTimeTableCreationHistory = async () => {
-  const response = await get<BaseResponse<TimeTableCreationHistory>>(
-    END_POINT.FETCH_TIMETABLE_CREATION_HISTORY,
-  );
-  return response.data;
 };
